♻️ #67 refactor: virtual 로직 분리
